Feature(static pages):faq and contact page design
