+++ packages/marketplace/src/index.tsx
@@ -1,4 +1,2 @@
 export { Button } from "./components/atoms/button";
 export { Alert } from "./components/atoms/alert";
-export { Input } from "./components/atoms/input";
-export { Textarea } from "./components/atoms/textarea";
